Pre-cache the next image and introduce "g"o command.
My file server can take up to two seconds to fetch the data for an
image. After displaying an image, read the next image to make sure it
is in the local machine's cache. This is a kludge, but a kludge that
seems to work well for me.

Introduce the goto_image command ('g') to jump to a particular image
in the current set.